Share one install-state claim ledger with Slack
The access-review connectors are gaining an app-install ceremony whose
callback is a public GET carrying a vendor proof that stays valid after
use, unlike an OAuth code the vendor burns itself. Nothing else stops a
browser refresh, a back-button or a link-preview prefetch from binding a
second connector, so the ceremony needs a single-use ledger.

Slack's install already had exactly that ledger, and the connector one
began as a verbatim copy of it: same six columns, same Claim/Complete/
Release, same stale-reclaim predicate, differing only in the table name.
Rather than keep two, this lifts the mechanism into one InstallStateClaim
parameterized by the table it writes to, with a constructor per ceremony.
The tables stay separate so a retention sweep over one never touches the
other's rows.

Folding them also collapses the stale-claim window, which had drifted to
five minutes for Slack and two for connectors with no stated reason for
either. It is now one constant, documented against the invariant that
actually constrains it: it must outlast the bounded network work a
ceremony does while holding a claim, and both are capped at 30s.

The test moves with it and now runs against both ledgers rather than
whichever one a change happened to touch. It also gains the tenant and
organization guards, which had none: the ledger is keyed on the state
digest alone, so those two predicates in the ON CONFLICT clause are all
that stop a second tenant replaying a leaked state from taking the claim
out from under its owner. Both were verified load-bearing by removing
them one at a time and watching the test fail.
